requirements.txt
discord.py
jishaku
trello
